Move render into ComponentParser
